refactor(authentication-directory): Added an Authentication directory to the pages directory and moved related pages accordingly
